Fix scrolling behavior with simple Gridview
way more direct, definitely should have started with this.

Just need to be sure to map row data so that it lines up with
the number of columns, which is easy by just setting the column
count to the length of the headers list.

Only thing to be aware of is that gridview fits all of its children
into its container size based on childAspectRatio, so if we want
the grid to scroll offscreen, we need to make its width larger than
screen size. This is easily accomplished by just specifying width
based on column count * column width

still need to lock the header row/first column