Sort seasons by year and sesion and fix some calculation bugs
show latest season first and oldest season last

fix comparison calculations that accidentally had addition instead
of multiplication for some stats

fix all time stat calculation that was altering the original player
object, causing the first season added to the list to display all
time stats instead of that seasons stats.

Instead, we need to create a new football player object that has
the correct totals when we add it to the all time array, instead of
altering the entry directly, since the original entry is a direct
reference to a player from the first season in the list.

There are no built-in ways to deep clone an object in dart unfortunately